Moved provideLocalStorageDataClass from utils to data
